move prefetch sounds to fix hot reload
