+++ src/useAudioContext.tsx
@@ -13,8 +13,7 @@ import {
   useSetting,
 } from "./settings/settings";
 import { useMediaDevices } from "./livekit/MediaDevicesContext";
-
-type SoundDefinition = { mp3?: string; ogg: string };
+import { PrefetchedSounds } from "./soundUtils";
 
 /**
  * Play a sound though a given AudioContext. Will take
@@ -37,58 +36,12 @@ function playSound(
   src.start();
 }
 
-/**
- * Determine the best format we can use to play our sounds
- * through. We prefer ogg support if possible, but will fall
- * back to MP3.
- * @returns "ogg" if the browser is likely to support it, or "mp3" otherwise.
- */
-function getPreferredAudioFormat(): "ogg" | "mp3" {
-  const a = document.createElement("audio");
-  if (a.canPlayType("audio/ogg") === "maybe") {
-    return "ogg";
-  }
-  // Otherwise just assume MP3, as that has a chance of being more widely supported.
-  return "mp3";
-}
-
-const preferredFormat = getPreferredAudioFormat();
-
-type PrefetchedSounds<S extends string> = Promise<Record<S, ArrayBuffer>>;
-
-/**
- * Prefetch sounds to be used by the AudioContext. This should
- * be called outside the scope of a component to ensure the
- * sounds load ahead of time.
- * @param sounds A set of sound files that may be played.
- * @returns A map of sound files to buffers.
- */
-export async function prefetchSounds<S extends string>(
-  sounds: Record<S, SoundDefinition>,
-): PrefetchedSounds<S> {
-  const buffers: Record<string, ArrayBuffer> = {};
-  await Promise.all(
-    Object.entries(sounds).map(async ([name, file]) => {
-      const { mp3, ogg } = file as SoundDefinition;
-      // Use preferred format, fallback to ogg if no mp3 is provided.
-      // Load an audio file
-      const response = await fetch(
-        preferredFormat === "ogg" ? ogg : (mp3 ?? ogg),
-      );
-      if (!response.ok) {
-        // If the sound doesn't load, it's not the end of the world. We won't play
-        // the sound when requested, but it's better than failing the whole application.
-        logger.warn(`Could not load sound ${name}, resposne was not okay`);
-        return;
-      }
-      // Decode it
-      buffers[name] = await response.arrayBuffer();
-    }),
-  );
-  return buffers as Record<S, ArrayBuffer>;
-}
-
 interface Props<S extends string> {
+  /**
+   * The sounds to play. If no sounds should be played then
+   * this can be set to null, which will prevent the audio
+   * context from being created.
+   */
   sounds: PrefetchedSounds<S> | null;
   latencyHint: AudioContextLatencyCategory;
 }
@@ -112,8 +65,8 @@ export function useAudioContext<S extends string>(
   const [audioBuffers, setAudioBuffers] = useState<Record<S, AudioBuffer>>();
 
   useEffect(() => {
-    const soundList = props.sounds;
-    if (!soundList) {
+    const sounds = props.sounds;
+    if (!sounds) {
       return;
     }
     const ctx = new AudioContext({
@@ -126,9 +79,7 @@ export function useAudioContext<S extends string>(
     // close during this process, so it's okay if it throws.
     (async (): Promise<void> => {
       const buffers: Record<string, AudioBuffer> = {};
-      for (const [name, buffer] of Object.entries<ArrayBuffer>(
-        await soundList,
-      )) {
+      for (const [name, buffer] of Object.entries<ArrayBuffer>(await sounds)) {
         const audioBuffer = await ctx.decodeAudioData(buffer.slice(0));
         buffers[name] = audioBuffer;
       }
